+++ src/translations/nl/easy-search.php
@@ -9,6 +9,8 @@ return [
     'is not equal to' => 'is niet gelijk aan',
     'is empty' => 'is leeg',
     'is not empty' => 'is niet leeg',
+    'is true' => 'is waar',
+    'is false' => 'is onwaar',
     'Add a condition' => 'Conditie toevoegen',
     'and' => 'en',
     'or' => 'of',
@@ -18,4 +20,7 @@ return [
     'Select a search query' => 'Selecteer een zoekopdracht',
     '... or build one' => '... of maak er één',
     'remove' => 'verwijder',
+    'Remove this condition' => 'Verwijde deze conditie',
+    'Show the results for this query' => 'Toon de resultaten voor deze zoekopdracht',
+    'Show results' => 'Toon resultaten',
 ];