+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Easy Search Changelog
 
+## 1.0.2 - 2021-04-20
+### Added
+- Added the `true` and `false` operators (only visible for lightswitch fields)
+
+### Fixed
+- Fixed a bug that caused fields to be shown multiple times if added to more than one entry type
+
+### Improved
+- Improved the way input fields respond to a user pressing `[enter]` (this now shows the results and closes the search hub)
+
 ## 1.0.1 - 2021-03-10
 ### Added
 - Added the default Commerce order fields to the list of possible fields to search for